+++ front-end/app/dashboard/locations/LocationsContent.tsx
@@ -3,13 +3,13 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 "use client"
 
+import Container from "@/components/Container"
 import Heading from "@/components/Heading"
 import ListingCard from "@/components/ListingCard"
 import AddListingModal from "@/components/modals/AddListingModal"
 import { useRouter } from "next/navigation"
 import { useCallback, useEffect, useState } from "react"
 import { toast } from "react-hot-toast"
-import Container from "@/components/Container";
 
 export default function LocationsContent() {
   const router = useRouter()
@@ -113,7 +113,7 @@ export default function LocationsContent() {
           setIsModalOpen(open)
           if (!open) setSelectedListing(null)
         }}
-        listing={selectedListing}
+        listingToEdit={selectedListing}
       />
 
       <Heading
