+++ front-end/components/modals/AddListingModal.tsx
@@ -1,58 +1,91 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+/* eslint-disable react/no-unescaped-entities */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 "use client"
 
-import { Button } from "@/components/ui/button"
-import { Checkbox } from "@/components/ui/checkbox"
+import { Button } from "@/components/ui/button";
+import { Checkbox } from "@/components/ui/checkbox";
 import {
   Dialog,
   DialogContent,
-  DialogDescription,
   DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
+  DialogTitle
+} from "@/components/ui/dialog";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "@/components/ui/select"
-import { Textarea } from "@/components/ui/textarea"
-import axios from "axios"
-import { useEffect, useState } from "react"
-import { toast } from "react-hot-toast"
+} from "@/components/ui/select";
+import { Textarea } from "@/components/ui/textarea";
+import axios from "axios";
+import { useRouter } from "next/navigation";
+import { useEffect, useState } from "react";
+import { toast } from "react-hot-toast";
 
-export default function ModalAjoutAnnonce({ open, onOpenChange }: any) {
+export default function AddListingModal({ open, onOpenChange, listingToEdit }: any) {
+  const router = useRouter();
   const [mounted, setMounted] = useState(false)
-  const [typeLogement, setTypeLogement] = useState<"maison" | "appartement">(
-    "maison"
-  )
   const [isLoading, setIsLoading] = useState(false)
   const [imageFile, setImageFile] = useState<File | null>(null)
+
   const [categories, setCategories] = useState<any[]>([])
   const [availableOptions, setAvailableOptions] = useState<any[]>([])
+  const [localisations, setLocalisations] = useState<any[]>([])
+
+  const [typeLogement, setTypeLogement] = useState<"maison" | "appartement">("maison")
+  const [title, setTitle] = useState("")
+  const [description, setDescription] = useState("")
+  const [selectedCity, setSelectedCity] = useState<string>("")
+  const [selectedCategory, setSelectedCategory] = useState<string>("")
+  const [price, setPrice] = useState<string>("")
+  const [capacity, setCapacity] = useState<string>("")
+  const [selectedOptions, setSelectedOptions] = useState<string[]>([])
+  const [manualImageUrl, setManualImageUrl] = useState("")
 
   const API_URL = process.env.NEXT_PUBLIC_API_URL || "https://localhost:8000"
+  const isEditMode = !!listingToEdit
+
+  useEffect(() => { setMounted(true) }, [])
 
   useEffect(() => {
-    setMounted(true)
-  }, [])
+    if (listingToEdit && open) {
+      setTypeLogement(listingToEdit.gardenSize !== undefined ? "maison" : "appartement")
+      setTitle(listingToEdit.title || "")
+      setDescription(listingToEdit.description || "")
+      setSelectedCity(listingToEdit.localisation?.id?.toString() || "")
+      setSelectedCategory(listingToEdit.category?.id?.toString() || "")
+      setPrice(listingToEdit.pricePerNight?.toString() || "")
+      setCapacity(listingToEdit.capacity?.toString() || "")
+      setSelectedOptions(listingToEdit.options?.map((o: any) => o.id.toString()) || [])
+      setManualImageUrl(listingToEdit.images?.[0]?.url || "")
+    } else if (open) {
+      setTitle(""); setDescription(""); setSelectedCity(""); setSelectedCategory("");
+      setPrice(""); setCapacity(""); setTypeLogement("maison"); setSelectedOptions([]);
+      setManualImageUrl(""); setImageFile(null);
+    }
+  }, [listingToEdit, open])
 
   useEffect(() => {
     if (open && mounted) {
       const fetchData = async () => {
         try {
-          const [resCats, resOpts] = await Promise.all([
-            fetch(`${API_URL}/api/categories`).then((r) => r.json()),
-            fetch(`${API_URL}/api/options`).then((r) => r.json()),
+          const token = localStorage.getItem("jwtToken")
+          const config = { headers: { 'Accept': 'application/ld+json', 'Authorization': `Bearer ${token}` } }
+          const [resCats, resOpts, resLocs] = await Promise.all([
+            axios.get(`${API_URL}/api/categories`, config),
+            axios.get(`${API_URL}/api/options`, config),
+            axios.get(`${API_URL}/api/localisations`, config)
           ])
-          setCategories(resCats.member || [])
-          setAvailableOptions(resOpts.member || [])
-        } catch (err) {
-          console.error("Erreur de chargement des données API", err)
+          setCategories(resCats.data["hydra:member"] || resCats.data.member || [])
+          setAvailableOptions(resOpts.data["hydra:member"] || resOpts.data.member || [])
+          setLocalisations(resLocs.data["hydra:member"] || resLocs.data.member || [])
+        } catch (error) {
+            toast.error("Erreur lors de la récupération des options du formulaire");
         }
       }
       fetchData()
@@ -60,80 +93,85 @@ export default function ModalAjoutAnnonce({ open, onOpenChange }: any) {
   }, [open, mounted, API_URL])
 
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault()
-    setIsLoading(true)
-    const formData = new FormData(e.currentTarget)
-    const token = localStorage.getItem("jwtToken")
-    console.log("Clique bouton envoie")
+    e.preventDefault();
+    setIsLoading(true);
+
+    const token = localStorage.getItem("jwtToken");
+    const formData = new FormData(e.currentTarget);
 
     try {
-      // 1. Récupération de l'URL saisie manuellement
-      let finalImageUrl = formData.get("imageUrl") as string
+      let imageToSubmit = manualImageUrl.trim();
 
-      console.log("Lien de l'image :", finalImageUrl)
-
-      // 2. Gestion de l'upload fichier (Prioritaire si présent)
       if (imageFile) {
-        const imgData = new FormData()
-        imgData.append("file", imageFile)
+        const imgData = new FormData();
+        imgData.append("file", imageFile);
         const up = await axios.post(`${API_URL}/api/upload-image`, imgData, {
-          headers: { Authorization: `Bearer ${token}` },
-        })
-        // Construction de l'URL absolue pour Next/Image
-        finalImageUrl = `${API_URL}${up.data.url}`
+          headers: { Authorization: `Bearer ${token}` }
+        });
+        imageToSubmit = up.data.url.startsWith('http') ? up.data.url : `${API_URL}${up.data.url}`;
       }
-
-      console.log("url finale", finalImageUrl)
-
-      // 3. Image de secours (Placeholder neutre si aucun des deux n'est rempli)
-      const fallback =
-        "https://images.unsplash.com/photo-1518780664697-55e3ad937233?q=80&w=1000"
 
       const payload: any = {
-        title: formData.get("title"),
-        description: formData.get("description"),
-        pricePerNight: Number(formData.get("pricePerNight")),
-        capacity: Number(formData.get("capacity")),
-        category: `/api/categories/${formData.get("category")}`,
-        options: formData.getAll("options[]").map((id) => `/api/options/${id}`),
-        images: [{ url: finalImageUrl || fallback }],
-      }
+        title,
+        description,
+        pricePerNight: parseFloat(price),
+        capacity: parseInt(capacity),
+        category: `/api/categories/${selectedCategory}`,
+        localisation: `/api/localisations/${selectedCity}`,
+        options: selectedOptions.map(id => `/api/options/${id}`),
+        images: imageToSubmit ? [{ url: imageToSubmit }] : []
+      };
 
-      // Logique Héritage
       if (typeLogement === "maison") {
-        payload.type = "house"
-        payload.gardenSize = Number(formData.get("gardenSize") || 0)
-        payload.hasGarage = formData.get("hasGarage") === "oui"
+        payload.gardenSize = Number(formData.get("gardenSize") || 0);
+        payload.hasGarage = formData.get("hasGarage") === "oui";
+        payload.type = "house";
       } else {
-        payload.type = "apartment"
-        payload.numberOfRooms = Number(formData.get("numberOfRooms") || 1)
-        payload.balcony = formData.get("balcony") === "oui"
+        payload.numberOfRooms = Number(formData.get("numberOfRooms") || 1);
+        payload.balcony = formData.get("balcony") === "oui";
+        payload.type = "apartment";
       }
 
-      await axios.post(`${API_URL}/api/listings`, payload, {
+      if (payload.images.length === 0 || payload.options.length === 0) {
+        toast.error("Veuillez ajouter au moins une image et un équipement.");
+        setIsLoading(false);
+        return;
+      }
+
+      const config = {
         headers: {
-          "Content-Type": "application/ld+json",
-          Authorization: `Bearer ${token}`,
-        },
-      })
+          "Authorization": `Bearer ${token}`,
+          "Accept": "application/ld+json",
+          "Content-Type": isEditMode ? "application/merge-patch+json" : "application/ld+json"
+        }
+      };
 
-      // ✅ NOTIFICATION DE SUCCÈS
-      toast.success("Annonce créée avec succès !")
+      if (isEditMode) {
+        await axios.patch(`${API_URL}/api/listings/${listingToEdit.id}`, payload, config);
+        toast.success("Annonce mise à jour !");
+      } else {
+        await axios.post(`${API_URL}/api/listings`, payload, config);
+        toast.success("Annonce publiée avec succès !");
+      }
 
-      // On ferme la modale avant le rechargement
-      if (onOpenChange) onOpenChange(false)
+      onOpenChange(false);
+      router.push("/dashboard#locations");
 
-      // Petit délai pour laisser le temps au toast d'être vu
       setTimeout(() => {
-        window.location.reload()
-      }, 1000)
+        window.location.reload();
+      }, 500);
+
     } catch (err: any) {
-      console.error(err)
-      toast.error("Erreur lors de la création de l'annonce")
+      const violations = err.response?.data?.violations;
+      if (violations) {
+        violations.forEach((v: any) => toast.error(v.message));
+      } else {
+        toast.error("Une erreur est survenue lors de l'enregistrement.");
+      }
     } finally {
-      setIsLoading(false)
+      setIsLoading(false);
     }
-  }
+  };
 
   if (!mounted) return null
 
@@ -141,78 +179,43 @@ export default function ModalAjoutAnnonce({ open, onOpenChange }: any) {
     <Dialog open={open} onOpenChange={onOpenChange}>
       <DialogContent className="max-w-2xl max-h-[90vh] overflow-y-auto">
         <DialogHeader>
-          <DialogTitle>Mettre mon logement en ligne</DialogTitle>
-          <DialogDescription>
-            Remplissez les détails de votre annonce pour la publier sur la
-            plateforme.
-          </DialogDescription>
+          <DialogTitle>{isEditMode ? "Modifier l'annonce" : "Publier une nouvelle annonce"}</DialogTitle>
         </DialogHeader>
 
         <form className="space-y-6" onSubmit={handleSubmit}>
-          <div className="space-y-3">
-            <Label className="font-bold">Type de bien</Label>
-            <RadioGroup
-              defaultValue="maison"
-              onValueChange={(v) => setTypeLogement(v as any)}
-              className="flex gap-4"
-            >
-              <div className="flex items-center space-x-2 border p-3 rounded-lg flex-1">
-                <RadioGroupItem value="maison" id="type-maison" />
-                <Label htmlFor="type-maison" className="cursor-pointer">
-                  Maison
-                </Label>
-              </div>
-              <div className="flex items-center space-x-2 border p-3 rounded-lg flex-1">
-                <RadioGroupItem value="appartement" id="type-appartement" />
-                <Label htmlFor="type-appartement" className="cursor-pointer">
-                  Appartement
-                </Label>
-              </div>
-            </RadioGroup>
-          </div>
+          {/* TYPE BIEN */}
+          <RadioGroup value={typeLogement} onValueChange={(v) => setTypeLogement(v as any)} className="flex gap-4">
+            <div className="flex items-center space-x-2 border p-3 rounded-lg flex-1 cursor-pointer">
+              <RadioGroupItem value="maison" id="m" />
+              <Label htmlFor="m" className="cursor-pointer">Maison</Label>
+            </div>
+            <div className="flex items-center space-x-2 border p-3 rounded-lg flex-1 cursor-pointer">
+              <RadioGroupItem value="appartement" id="a" />
+              <Label htmlFor="a" className="cursor-pointer">Appartement</Label>
+            </div>
+          </RadioGroup>
 
+          {/* TITRE & DESC */}
           <div className="grid gap-4">
-            <div className="grid gap-2">
-              <Label htmlFor="title">Titre</Label>
-              <Input
-                name="title"
-                id="title"
-                placeholder="Ex: Villa bord de mer"
-                required
-              />
-            </div>
-            <div className="grid gap-2">
-              <Label htmlFor="description">Description</Label>
-              <Textarea
-                name="description"
-                id="description"
-                placeholder="Décrivez votre bien..."
-                required
-              />
-            </div>
+            <Input value={title} onChange={(e) => setTitle(e.target.value)} placeholder="Titre de l'annonce *" required />
+            <Textarea value={description} onChange={(e) => setDescription(e.target.value)} placeholder="Description détaillée *" required />
           </div>
 
-          <div className="grid grid-cols-2 gap-4 bg-gray-50 p-4 rounded-lg border">
+          {/* SPECIFIQUES */}
+          <div className="grid grid-cols-2 gap-4 bg-muted/50 p-4 rounded-lg border">
             {typeLogement === "maison" ? (
               <>
                 <div className="grid gap-2">
-                  <Label htmlFor="gardenSize">Surface jardin (m²)</Label>
-                  <Input
-                    name="gardenSize"
-                    id="gardenSize"
-                    type="number"
-                    defaultValue={0}
-                  />
+                    <Label>Jardin (m²)</Label>
+                    <Input name="gardenSize" type="number" defaultValue={listingToEdit?.gardenSize || 0} />
                 </div>
                 <div className="grid gap-2">
                   <Label>Garage ?</Label>
-                  <Select name="hasGarage" defaultValue="non">
-                    <SelectTrigger>
-                      <SelectValue placeholder="Choisir" />
-                    </SelectTrigger>
+                  <Select name="hasGarage" defaultValue={listingToEdit?.hasGarage ? "oui" : "non"}>
+                    <SelectTrigger><SelectValue /></SelectTrigger>
                     <SelectContent>
-                      <SelectItem value="oui">Oui</SelectItem>
-                      <SelectItem value="non">Non</SelectItem>
+                        <SelectItem value="oui">Oui</SelectItem>
+                        <SelectItem value="non">Non</SelectItem>
                     </SelectContent>
                   </Select>
                 </div>
@@ -220,24 +223,16 @@ export default function ModalAjoutAnnonce({ open, onOpenChange }: any) {
             ) : (
               <>
                 <div className="grid gap-2">
-                  <Label htmlFor="numberOfRooms">Nombre de pièces</Label>
-                  <Input
-                    name="numberOfRooms"
-                    id="numberOfRooms"
-                    type="number"
-                    defaultValue={1}
-                    required
-                  />
+                    <Label>Pièces</Label>
+                    <Input name="numberOfRooms" type="number" defaultValue={listingToEdit?.numberOfRooms || 1} required />
                 </div>
                 <div className="grid gap-2">
                   <Label>Balcon ?</Label>
-                  <Select name="balcony" defaultValue="non">
-                    <SelectTrigger>
-                      <SelectValue placeholder="Choisir" />
-                    </SelectTrigger>
+                  <Select name="balcony" defaultValue={listingToEdit?.balcony ? "oui" : "non"}>
+                    <SelectTrigger><SelectValue /></SelectTrigger>
                     <SelectContent>
-                      <SelectItem value="oui">Oui</SelectItem>
-                      <SelectItem value="non">Non</SelectItem>
+                        <SelectItem value="oui">Oui</SelectItem>
+                        <SelectItem value="non">Non</SelectItem>
                     </SelectContent>
                   </Select>
                 </div>
@@ -245,101 +240,67 @@ export default function ModalAjoutAnnonce({ open, onOpenChange }: any) {
             )}
           </div>
 
-          <div className="grid grid-cols-3 gap-4">
-            <div className="grid gap-2">
-              <Label>Prix / nuit (€)</Label>
-              <Input name="pricePerNight" type="number" required />
-            </div>
-            <div className="grid gap-2">
-              <Label>Capacité</Label>
-              <Input name="capacity" type="number" required />
-            </div>
-            <div className="grid gap-2">
-              <Label>Catégorie</Label>
-              <Select name="category" required key={`cat-${categories.length}`}>
-                <SelectTrigger>
-                  <SelectValue placeholder="Choisir" />
-                </SelectTrigger>
-                <SelectContent>
-                  {Array.isArray(categories) &&
-                    categories.map(
-                      (cat) =>
-                        cat?.id && (
-                          <SelectItem key={cat.id} value={cat.id.toString()}>
-                            {cat.name}
-                          </SelectItem>
-                        )
-                    )}
-                </SelectContent>
-              </Select>
-            </div>
+          {/* VILLE & CAT */}
+          <div className="grid grid-cols-2 gap-4">
+            <Select value={selectedCity} onValueChange={setSelectedCity} required>
+              <SelectTrigger><SelectValue placeholder="Ville *" /></SelectTrigger>
+              <SelectContent className="z-9999">
+                {localisations.map((loc) => <SelectItem key={loc.id} value={loc.id.toString()}>{loc.name}</SelectItem>)}
+              </SelectContent>
+            </Select>
+            <Select value={selectedCategory} onValueChange={setSelectedCategory} required>
+              <SelectTrigger><SelectValue placeholder="Catégorie *" /></SelectTrigger>
+              <SelectContent className="z-9999">
+                {categories.map((cat) => <SelectItem key={cat.id} value={cat.id.toString()}>{cat.name}</SelectItem>)}
+              </SelectContent>
+            </Select>
           </div>
 
+          {/* PRIX & CAPACITE */}
+          <div className="grid grid-cols-2 gap-4">
+            <Input type="number" value={price} onChange={(e) => setPrice(e.target.value)} placeholder="Prix / nuit *" required />
+            <Input type="number" value={capacity} onChange={(e) => setCapacity(e.target.value)} placeholder="Capacité *" required />
+          </div>
+
+          {/* EQUIPEMENTS */}
           <div className="grid gap-2 border p-4 rounded-xl">
-            <Label className="font-bold text-xs uppercase opacity-50">
-              Options & Équipements
-            </Label>
+            <Label className="font-bold text-xs uppercase text-muted-foreground mb-2">Équipements (requis) *</Label>
             <div className="grid grid-cols-2 gap-3">
-              {Array.isArray(availableOptions) &&
-                availableOptions.map(
-                  (opt) =>
-                    opt?.id && (
-                      <div key={opt.id} className="flex items-center space-x-2">
-                        <Checkbox
-                          name="options[]"
-                          value={opt.id.toString()}
-                          id={`opt-${opt.id}`}
-                        />
-                        <Label
-                          htmlFor={`opt-${opt.id}`}
-                          className="cursor-pointer text-sm"
-                        >
-                          {opt.name}
-                        </Label>
-                      </div>
-                    )
-                )}
+              {availableOptions.map((opt) => (
+                <div key={opt.id} className="flex items-center space-x-2">
+                  <Checkbox
+                    id={`opt-${opt.id}`}
+                    checked={selectedOptions.includes(opt.id.toString())}
+                    onCheckedChange={(checked) => {
+                      if (checked) setSelectedOptions(prev => [...prev, opt.id.toString()])
+                      else setSelectedOptions(prev => prev.filter(id => id !== opt.id.toString()))
+                    }}
+                  />
+                  <Label htmlFor={`opt-${opt.id}`} className="cursor-pointer text-sm">{opt.name}</Label>
+                </div>
+              ))}
             </div>
           </div>
 
-          <div className="border-t pt-6 space-y-4">
-            <Label className="font-bold">Photos</Label>
-            <div className="grid gap-2">
-              <Label className="text-xs text-gray-500">
-                Uploader un fichier
-              </Label>
-              <Input
-                type="file"
-                accept="image/*"
-                onChange={(e) => setImageFile(e.target.files?.[0] || null)}
-              />
-            </div>
-            <div className="grid gap-2">
-              <Label className="text-xs text-gray-500">
-                Ou lien URL direct (Unsplash, Pexels...)
-              </Label>
-              <Input
-                name="imageUrl"
-                type="url"
-                placeholder="https://images.unsplash.com/..."
-              />
+          {/* PHOTOS */}
+          <div className="border-t pt-4 space-y-4">
+            <Label className="font-bold">Photos *</Label>
+            <div className="grid gap-3">
+                <Input type="file" accept="image/*" onChange={(e) => setImageFile(e.target.files?.[0] || null)} />
+                <div className="text-center text-xs text-muted-foreground">ou URL de l'image :</div>
+                <Input
+                    value={manualImageUrl}
+                    onChange={(e) => setManualImageUrl(e.target.value)}
+                    type="url"
+                    placeholder="Lien URL (ex: Unsplash)"
+                />
             </div>
           </div>
 
-          <div className="flex gap-3 justify-end pt-4 border-t">
-            <Button
-              type="button"
-              variant="outline"
-              onClick={() => onOpenChange(false)}
-            >
-              Annuler
-            </Button>
-            <Button
-              type="submit"
-              disabled={isLoading || categories.length === 0}
-              className="bg-rose-600 text-white px-10"
-            >
-              {isLoading ? "Envoi en cours..." : "Publier l'annonce"}
+          <div className="flex gap-3 justify-end pt-6 border-t">
+            <Button type="button" variant="ghost" onClick={() => onOpenChange(false)}>Annuler</Button>
+            <Button type="submit" disabled={isLoading} className="bg-green-600 hover:bg-green-700 text-white min-w-[140px]">
+              {isLoading ? "Envoi en cours..." : "Valider"}
             </Button>
           </div>
         </form>
